webapp/requirements.txt
torch
torchvision
numpy
pillow
uvicorn
fastapi
aiofiles
python-multipart
skimage